requirements.txt
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.40.2
transformers==4.31.0
trl==0.4.7
scipy==1.12.0
jmespath==1.0.1
fastapi==0.68.0
uvicorn==0.14.0
python-multipart==0.0.5
protobuf3==0.2.1
grpcio-tools==1.62.1